Use hostname instead of host
It is very unlikely that there is a case where we would want
to distinguish hosts on the basis of ports, to determine which
rules we want to apply. Hostname doesn't include the portnumber,
so we pick that one. It is cleaner and more reliable (e.g.
some-site.com:8080 will also trigger the rules for some-site.com),
and helps with testing, since we can now navigate to
'my-fake-url.com:9999', and let a server listen on that (and check
that all rules for 'my-fake-url.com' are applied). This has
the advantage that we don't need root permissions to listen on lower
(e.g. :80) ports. (we only need to modify /etc/hosts once with sudo)