Move test rules to seperate folder
Quite a major restructuring in how we load / use
test files: way better, since we are now independent
from external websites (e.g. we used example.com).
And we now don't ship test rulefiles in the production
build.
This is in preperation for the background functional
tests, since there it is important to have control over
the precise content of the sites we're serving.