Fix incorrect test for applyStatusChange
Boy, this took me a while. The main problem was
that applyStatusChange is async, which I hadn't
taken into account. Therefore, the assertion was
run before applyStatusChange had done its work.
Stupid error of me.

I also made some mistakes for the tabId: I didn't
resolve the browser.tabs.query-mock with the variable,
but with a hand coded integer.

What made this even more frustrating to debug, was a bug
in Sinon I found through this: comparing '1234' to 1234
gave a very confusing AssertionError message, as if
everything was okay. See https://github.com/sinonjs/sinon/issues/2084
for the bug I filed.